Revert "Add Tile object creation"
Gameboards do not need to know about the concept of a tile as it
overcomplicated the code and testing. A better approach will be to
have two boards within a gameboard object. One tracks ship placement,
and one tracks the player's shots. This is similar to how real life
battleship is, as a single player has two boards: one containing their
ships, and another containing their shots.

It might be worth utilizing the states enum concept for the shot board,
as this keeps the ideas separate.