Add basic site layout and grid generation
With the work on the backend finally done (aka creating the models to
represent out players and gameboard), work on the frontend can finally
start. This patch creates a fairly generic page that includes a header,
and a footer linking to my github.

Additionally grid generation was added in index.js. Upon the page
loading we query both grid elements and add our cells to them. The cell
elements contain data attributes that correspond to their row and column
indices. The attributes are 0 indexed so this should be fairly easy for
working with the backend objects.