Add controller and view object creation
The controller and view are responsible for logic pertaining to user
input and logic for UI respectively. These modules will help connect to
our data models and get out gameplay loop up and running.

As of this patch the game is playable, however it's bare-bones. There's
more functionality and styles to be added in future commits.