Add removeShip functionality to gameboard model
While working on the controller it was realized there should be a way to
remove ships from the model. This will allow players to have their ships
repositioned.

To keep track of where the ships are located for removal purposes, we
introduced an object that tracks shipData, specifically it's used for
the initial placement information. With that information we can figure
out where to remove the ship.

Lastly we introduced the concept of a shipType. These types correlate
with the ship's name. This means we can look up ship data using their
name, as there should only ever be once instance of a ship of a given
type.