Add game over status message and restart functionality
Adds functionality to the view and controller to create a 'Play Again'
button that resets the models and view to match the initial state.

Previously players would have to refresh the page to play again, this
allows players to restart without that poor UX experience.

Additionally the status message now properly informs the player of the
winner instead of utilizing the alert function.