Refactor handleReceiveAttack in controller
This was written in a way that was intended for only a player vs a bot.
Keeping in mind that I'd like to eventually allow for a two player mode,
this couldn't remain. Along with that it was a little hard to read and
had repeated sections.

This patch isolates the logic that was contained within
handleReceiveAttack into a separate concise function, playerTurn, that
goes through the logic for a single player's turn.