Update computer's intelligence
The previous computer intelligence was very primitive, randomly choosing
a valid tile and attacking it. Now the computer uses an extended version
of the player object with a method to get it's next attack. This is
attack is random until it hits a ship. The next attack will then be an
adjacent tile, otherwise if the previous attack was a miss it will be go
back to the default random behavior.