Refactor getInitialPosition to use shipType
Previously getInitialPosition took the row and col, which internally
got the ship then the type. This became a problem as sometimes the row
and col of the ship data we want isn't available.

This patch refactors the function to instead take in the ship's type as
an argument. This removes the internal intermediate step of getting the
ship with the coordinates, and makes more sense.