Add drag and drop ship placement system
This is one of the foundational features for the UI. Previously users
would have to randomize ship placement for their boards. Using the HTML
drag and drop API, we have set up event listeners for drag events. This
enables ships to be picked up and placed in a valid location on the
board.