Add visual feedback to ship drag and drop
This is the final part of the drag and drop implementation. This patch
focuses on missing UI/UX for the drag and drop placement system, adding
a feedback image and a insertion marker that changes color depending on
the validity of a placement. These are displayed using the events
provided by the drag and drop API.

Note: Some of the changes done here are done to fix functionality for
chrome. There is a known bug where the dragend event will fire
prematurely if the DOM is manipulated during a dragstart event.