Add form validations to planning modal
The planning modal was using the built-in validations provided by HTML
form controls. We want the validation messages and styles to be
consistent across browsers rather than using the default styles. This
patch adds custom form validations for that, including validations for
the planning board which cannot use the constraint validation API.

Additionally we've added the styles for these validations to help guide
users when planning their ship placement.