Refactor view methods to be more abstract
Rather than relying on the controller to know how to render the game at
a moment in time, we should abstract this into something the view
handles. This patch does exactly that, taking areas of the controller
that called multiple view methods to render part of the game, and
abstracting them to view methods. Now the controller only has to call
the corresponding render method.