Refactor binding event listeners to handlers
Previously the logic for where we bound handlers to event listeners was
inconsistent. Sometimes checks were done in the controller, other times
in the view. This was not thought out and often done to handle event
delegation within a single handler. This patch changes that so event
delegation logic is contained within the view, breaking up the old
handlers into separate concise handlers.