Improve computer attacking logic
Previously the computer used only the last attack in deciding which tile
to attack. However this caused it to not attack ships further upon the
first miss. To improve this, we've replaced the object storing the last
attack with a stack data structure that pushes adjacent tiles onto the
stack. Additionally there is logic to continue attacking along a certain
plane if there are multiple hits in a row.