Clean up element styles
There were a couple of adjustments I'd been meaning to make in the
styles file. The main part of this change was moving inline styles being
applied in JS to styles applied to a class, if applicable. Some elements
such as grid cells weren't applicable due to how we use overlap with
dynamic grid areas.